check {rqPen}	R Documentation

Quantile check function

Description

Evaluates the check function for quantile tau at value x. Check function is the objective function defined in Koenker and Bassett (1978).

Usage

check(x, tau)

Arguments

x
Number to be evaluated.

tau
Number between 0 and 1 for quantile of interest.

Value

x*(tau-I(x < 0)

Author(s)

Ben Sherwood

References

[1] Koenker, R. and Bassett, G. (1978). Regression Quantiles, Econometrica, 46, 33–50.

Examples

check(2,.5)
check(-2,.5)
check(2,.2)
check(2,.8)

[Package rqPen version 1.4 Index]