coef.cv.rq.pen {rqPen}	R Documentation

Penalized Quantile Regression Coefficients

Description

Returns coefficients for a cv.rq.pen object with default values being coefficients associated with minimum cross-validation value.

Usage

## S3 method for class 'cv.rq.pen'
coef(object, lambda='min',...)

Arguments

object
cv.rq.pen object

lambda
Tuning parameter lambda. Default is to select minimum lambda from cross-validation method. User can also select a specific value of lambda, but it needs to be a lambda that was part of the fit of cv.rq.pen object.

...
Additional arguments, currently not used.

Value

Coefficients for selected value of lambda.

Author(s)

Ben Sherwood

Examples

x <- matrix(rnorm(800),nrow=100)
y <- 1 + x[,1] - 3*x[,5] + rnorm(100)
cv_model <- cv.rq.pen(x,y)
coefficients(cv_model)

[Package rqPen version 1.4 Index]