getRho {rqPen}	R Documentation

Objective Function Value

Description

Returns unpenalized portion of the objective function for a penalized quantile regression object.

Usage

getRho(model)

Arguments

model
Object with rho as an attribute

Value

Rho is the value of the unpenalized portion of the objective function.

Author(s)

Ben Sherwood

Examples

x <- matrix(rnorm(800),nrow=100)
y <- 1 + x[,1] - 3*x[,5] + rnorm(100)
lassoModel <- rq.lasso.fit(x,y,lambda=.5)
getRho(lassoModel)

[Package rqPen version 1.4 Index]