kernel_weights {rqPen}	R Documentation

Nonparametric estimate of IPW weights

Description

This is for downstream analysis for fitting models with missing data. Future work is to fully incorporate these into penalized models. Tuning parameter for condtional density is esimated using approach of Chen, Wan and Zhou (2015), which is a simplified approach of Sepanski et al. (1994)

Usage

kernel_weights(obs_data,obs_ind,...)

Arguments

obs_data
Matrix of variables with complete observations

obs_ind
Vector of whether sample is observed or not (1-observed, 0-not)

...
Additional arguments to be sent to kernel_estimates.

Value

Estimates of weights.

Author(s)

Ben Sherwood

References

[1] Chen, X., Wan, A. and Zhou, Y. Efficient quantile regression analysis with missing observations. (2015). J. Amer. Statist. Assoc., 110, 723–741. [2] Sepanski, J., Knickerbocker, R. and Carroll, R. A semiparametric correction for attenuation. (1994). J. Amer. Statist. Assoc., 89, 1366–1373.

Examples

x <- matrix(rnorm(800),nrow=100)
y <- rbinom(100,1,.5)
wt_est <- kernel_weights(x,y)

[Package rqPen version 1.4 Index]