qbic {rqPen}	R Documentation

Quantile Regresion BIC

Description

Quantile regression BIC with large p alternative as described in Lee, Noh and Park (2013).

Usage

qbic(model, largeP=FALSE)

Arguments

model
Model of class "rqPen".

largeP
Large P version using an additional penalty factor of log(s) where "s" is the total number of covariates considered.

Value

Numeric value representing BIC of selected model.

Author(s)

Ben Sherwood

References

[1] Lee, E., Noh, H. and Park, B. (2014). Model selection via Bayesian Information Criterion for quantile regression models., J. Am. Statist. Ass, 109, 216–229.

Examples

x <- matrix(rnorm(800),nrow=100)
y <- 1 + x[,1] - 3*x[,5] + rnorm(100)
l_model <- rq.lasso.fit(x,y, lambda=1)
nc_model <- rq.nc.fit(x,y, lambda=1)
qbic(l_model)
qbic(nc_model)
qbic(l_model, largeP=TRUE)
qbic(nc_model, largeP=TRUE)

[Package rqPen version 1.4 Index]