rq.group.fit {rqPen}	R Documentation

Quantile Regresion with Group Penalty

Description

Similar to cv.rq.pen function, but uses group penalty. Group penalties use the L1 norm instead of L2 for computational convenience. We use a group penalty extension of the QICD algorithm presented by Peng and Wang (2015).

Usage

rq.group.fit(x, y, groups, tau = 0.5, lambda, intercept = TRUE, 
                penalty = "LASSO", alg = "QICD", ...)

Arguments

x
Matrix of predictors.

y
Vector of response values.

groups
Vector assigning columns of x to groups.

tau
Conditional quantile being modelled.

lambda
Vector of lambdas. Default is for lambdas to be automatically generated.

intercept
Whether model should include an intercept. Constant does not need to be included in "x".

penalty
Type of penalty: "LASSO", "SCAD" or "MCP".

alg
If set to QICD algorithm will use coordinate descent algorithm. Otherwise, will use a linear programming algorithm, which is an extension of algorithm used in rq.lasso.fit.

...
Additional arguments to be sent to rq.lasso.fit or groupQICD.

Value

Returns the following:

coefficients
Coefficients of the model.

residuals
Residuals from the fitted model.

rho
Unpenalized portion of the objective function.

tau
Quantile being modelled.

n
Sample size.

intercept
Whether intercept was included in model.

Author(s)

Ben Sherwood

References

[1] Yuan, M. and Lin, Y. (2006). Model selection and estimation in regression with grouped variables. J. R. Statist. Soc. B, 68, 49-67.

[2] Peng, B. and Wang, L. (2015). An Iterative Coordinate Descent Algorithm for High-Dimensional Nonconvex Penalized Quantile Regression. Journal of Computational and Graphical Statistics, 24, 676-694.

Examples

x <- matrix(rnorm(800),nrow=100)
y <- 1 + x[,1] - 3*x[,5] + rnorm(100)
cv_model <- rq.group.fit(x,y,groups=c(rep(1,4),rep(2,4)),lambda=1)

[Package rqPen version 1.4 Index]