rq.lasso.fit.mult {rqPen}	R Documentation

Fit Quantile Regression model for varying quantiles with LASSO penalty

Description

Fits quantile regression models for multiple quantiles with the LASSO penalty. Algorithm is similar to LASSO code presented in Koenker and Mizera (2014).

Usage

rq.lasso.fit.mult(x,y,tau_seq=c(.1,.3,.5,.7,.9),lambda=NULL,
          weights=NULL,intercept=TRUE,coef.cutoff=.00000001,...)

Arguments

x
Matrix of predictors.

y
Vector of response values.

tau_seq
Vector of quantiles of interest

lambda
Tuning parameter.

weights
Weights for the objective function.

intercept
Whether model should include an intercept. Constant does not need to be included in "x".

coef.cutoff
Coefficients below this value will be set to zero.

...
Additional items to be sent to rq. Note this will have to be done carefully as rq is run on the augmented data to account for penalization and could provide strange results if this is not taken into account.

Value

Returns a list of rq.pen, rqLASSO objects.

Author(s)

Ben Sherwood

References

[1] Koenker, R. and Mizera, I. (2014). Convex optimization in R. Journal of Statistical Software, 60, 1–23.

[2] Tibshirani, R. (1996). Regression shrinkage and selection via the lasso. Journal of the Royal Statistical Society. Series B, 58, 267–288.

[3] Wu, Y. and Liu, Y. (2009). Variable selection in quantile regression. Statistica Sinica, 19, 801–817.

Examples

x <- matrix(rnorm(800),nrow=100)
y <- 1 + x[,1] - 3*x[,5] + rnorm(100)
lassoModel <- rq.lasso.fit.mult(x,y,lambda=1)

[Package rqPen version 1.4 Index]