scad_deriv {rqPen}	R Documentation

SCAD Derivative

Description

Derivative of SCAD penalty function as described in Fan and Li (2001).

Usage

scad_deriv(x,lambda,a)

Arguments

x
Number to be evaluated

lambda
Tuning parameter lambda

a
Tuning parameter a. Default value of 3.7 as suggested in Fan and Li (2001)

Value

Derivative SCAD penalty function with tuning parameters lambda and "a" evaluated at "x".

Author(s)

Ben Sherwood

References

[1] Fan, J. and Li, R. (2001). Variable selection via nonconcave penalized likelihood and its oracle properties, J. Am. Statist. Ass, 96, 1348–1360.

Examples

scad_deriv(3,1)
scad_deriv(-3,1)
scad_deriv(.001,2)

[Package rqPen version 1.4 Index]