cv_plots {rqPen}	R Documentation

Plots of Cross-validation results

Description

Slightly misnamed as user could choose BIC as a criteria for "cv.rq.pen" object. Function is able to discern between the two types of evaluation criteria and provides appropriate labels for the plot.

Usage

cv_plots(model,logLambda=TRUE,loi=NULL,...)

Arguments

model
"cv.rq.pen" object.

logLambda
Plot of lambdas is on the log scale.

loi
Index of lambdas to be plotted. Default is all of the lambdas from "cv.rq.pen" object.

...
Additional items to be sent to plot function.

Value

Plot of how cross validation statistic changes with lambda.

Author(s)

Ben Sherwood

Examples

x <- matrix(rnorm(800),nrow=100)
y <- 1 + x[,1] - 3*x[,5] + rnorm(100)
lassoModels <- cv.rq.pen(x,y)
cv_plot <- cv_plots(lassoModels)

[Package rqPen version 1.4 Index]